+++ jest.config.js
@@ -2,19 +2,18 @@
 module.exports = {
   preset: "react-native",
   verbose: true,
+  setupFiles: ["./node_modules/react-native-gesture-handler/jestSetup.js"],
   setupFilesAfterEnv: [
     "./jest.setup.js"
   ],
   testMatch: ["**/src/**/?(*.)+(spec|test).[jt]s?(x)"],
   transformIgnorePatterns: [
-    'node_modules/(?!(jest-)?react-native|@react-native|@react-native-community|@react-navigation|ky)',
+    'node_modules/(?!(jest-)?react-native|react-native-reanimated|@react-native|@react-native-community|@react-navigation|ky)',
+    "node_modules/(?!(@react-native|react-native|react-native-reanimated)/)"
   ],
   moduleFileExtensions: ['js', 'jsx', 'ts', 'tsx', 'json', 'node'],
   transform: {
     '^.+\\.(js|jsx|ts|tsx)$': 'babel-jest', // Ensure Babel is used for transforming JS files
-    // "^.+\\.ts?$": "ts-jest",
-    // "^.+\\.tsx?$": "ts-jest",
-    // "^.+\\.js$": "./node_modules/react-native/jest/preprocessor.js"
   },
   // transformOptions: {
   //   babelConfig: require('./babel.base.js'),
@@ -30,9 +29,15 @@ module.exports = {
   // moduleNameMapper: {
   //   '^@babel/runtime': '<rootDir>/node_modules/@babel/runtime',
   // },
-  moduleNameMapper: {
-    '^@react-native-ama/core$': '<rootDir>/packages/core/src',
-  },
+  // moduleNameMapper: {
+  //   '^@react-native-ama/core$': '<rootDir>/packages/core/src',
+  // },
+  //moduleDirectories: ['node_modules', 'packages'], // Added for module resolution
+  // moduleNameMapper: {
+  //   '^@react-native-ama/internal$': '<rootDir>/packages/core/src', // Adjust as necessary
+  //   '^@react-native-ama/core$': '<rootDir>/packages/core/src',
+  // },
+  testTimeout: 10000,
   projects: [
     {
       testPathIgnorePatterns: ["<rootDir>/node_modules/"],
@@ -48,6 +53,38 @@ module.exports = {
       displayName: "core",
       testMatch: ["<rootDir>/packages/core/src/**/*.test.ts",
           "<rootDir>/packages/core/src/**/*.test.tsx"
+      ],
+    },
+    {
+      testPathIgnorePatterns: ["<rootDir>/node_modules/"],
+      preset: "react-native",
+      displayName: "internal",
+      testMatch: ["<rootDir>/packages/internal/src/**/*.test.ts",
+          "<rootDir>/packages/internal/src/**/*.test.tsx"
+      ],
+    },
+    {
+      testPathIgnorePatterns: ["<rootDir>/node_modules/"],
+      preset: "react-native",
+      displayName: "extras",
+      testMatch: ["<rootDir>/packages/extras/src/**/*.test.ts",
+          "<rootDir>/packages/extras/src/**/*.test.tsx"
+      ],
+    },
+    {
+      testPathIgnorePatterns: ["<rootDir>/node_modules/"],
+      preset: "react-native",
+      displayName: "lists",
+      testMatch: ["<rootDir>/packages/lists/src/**/*.test.ts",
+          "<rootDir>/packages/lists/src/**/*.test.tsx"
+      ],
+    },
+    {
+      testPathIgnorePatterns: ["<rootDir>/node_modules/"],
+      preset: "react-native",
+      displayName: "forms",
+      testMatch: ["<rootDir>/packages/forms/src/**/*.test.ts",
+          "<rootDir>/packages/forms/src/**/*.test.tsx"
       ],
     }
   ],
