+++ jest.setup.js
@@ -1,6 +1,4 @@
-// import 'react-native-gesture-handler/jestSetup';
 
-// require('react-native-reanimated/lib/reanimated2/jestUtils').setUpTests();
 import '@react-native-community/jest-setup';
 import 'react-native-gesture-handler/jestSetup';
 require('react-native-reanimated/lib/reanimated2/jestUtils').setUpTests();
