+++ package.json
@@ -22,7 +22,12 @@
     "version": "yarn changeset version && yarn install --no-frozen-lockfile",
     "build": "yarn build:internal && yarn build:core && yarn build:animations && yarn build:react-native && yarn build:extras && yarn build:forms && yarn build:lists",
     "test:animations": "jest --selectProjects=animations",
-    "test:core": "jest --selectProjects=core"
+    "test:core": "jest --selectProjects=core",
+    "test:internal": "jest --selectProjects=internal",
+    "test:extras": "jest --selectProjects=extras",
+    "test:lists": "jest --selectProjects=lists",
+    "test:forms": "jest --selectProjects=forms",
+    "test": "jest --selectProjects=animations && jest --selectProjects=core && jest --selectProjects=internal && jest --selectProjects=extra"
   },
   "workspaces": [
     "packages/animations",
@@ -56,7 +61,8 @@
     "eslint-plugin-prettier": "^3.1.3",
     "husky": "^8.0.1",
     "jest": "^29.7.0",
-    "react-native-reanimated": "^3.15.5",
+    "react-native-reanimated": ">=2.0.0",
+    "react-native-gesture-handler": ">=2.0.0",
     "react-test-renderer": "^18.3.1",
     "ts-jest": "^29.1.1",
     "typescript": "^5.1.3"
